% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splice.R
\name{all_}
\alias{all_}
\alias{if_}
\alias{at_}
\title{Splice utilities}
\usage{
all_(.funs, ...)

if_(.predicate, .funs, ...)

at_(.vars, .funs, ...)
}
\arguments{
\item{.funs}{A function `fun`, a quosure style lambda `~ fun(.)` or a list of either form.}

\item{...}{Additional arguments for the function calls in
`.funs`. These are evaluated only once, with [tidy
dots][rlang::tidy-dots] support.}

\item{.predicate}{A predicate function to be applied to the columns
or a logical vector. The variables for which `.predicate` is or
returns `TRUE` are selected. This argument is passed to
[rlang::as_function()] and thus supports quosure-style lambda
functions and strings representing function names.}

\item{.vars}{A list of columns generated by [dplyr::vars()],
a character vector of column names, a numeric vector of column
positions, or `NULL`.}
}
\description{
The helpers `all_()`, `if_()` and `at_()` are meant to be spliced
into calls to `dplyr::summarise()` or `dplyr::mutate()`
}
\examples{
library(dplyr, warn.conflicts = FALSE)
tbl <- iris \%>\%
  group_by(Species)

# all_
tbl \%>\%
  summarise(n = n(), !!!all_(mean))

tbl \%>\%
  summarise(n = n(), !!!all_(~mean(.)))

tbl \%>\%
  summarise(n = n(), !!!all_(list(mean = mean, median = median)))

# if_
iris \%>\%
  summarise(n = n(), !!!if_(is.numeric, mean))

# at_
iris \%>\%
  summarise(n = n(), !!!at_(vars(starts_with("Sepal")), mean))
}
