% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_incurred.R
\name{generate_incurred_dataset}
\alias{generate_incurred_dataset}
\title{Generate Incurred Dataset}
\usage{
generate_incurred_dataset(claims, incurred_history)
}
\arguments{
\item{claims}{an \code{claims} object containing all the simulated quantities,
(other than those related to incurred loss), see
\code{\link[SynthETIC]{claims}}.}

\item{incurred_history}{the full history of incurred case estimates, see
\code{\link{claim_history}}.}
}
\value{
A dataframe that takes the same structure as
\code{\link{test_incurred_dataset}}.
}
\description{
Generates a dataset of transaction records that tracks how the case estimates
of the total incurred, the outstanding claim payments, and the cumulative
claims paid change over time. A sample dataset is included as part of the
package, see \code{\link{test_incurred_dataset}}.
}
\seealso{
\code{\link{test_incurred_dataset}}
}
