% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPLICE-package.R
\docType{package}
\name{SPLICE-package}
\alias{SPLICE}
\alias{SPLICE-package}
\title{SPLICE: Synthetic Paid Loss and Incurred Cost Experience (SPLICE) Simulator}
\description{
An extension to the individual claim simulator called 'SynthETIC' (on CRAN), to simulate the evolution of case estimates of incurred losses through the lifetime of an insurance claim. The transactional simulation output now comprises key dates, and both claim payments and revisions of estimated incurred losses. An initial set of test parameters, designed to mirror the experience of a real insurance portfolio, were set up and applied by default to generate a realistic test data set of incurred histories (see vignette). However, the distributional assumptions used to generate this data set can be easily modified by users to match their experiences. Reference: Avanzi B, Taylor G, Wang M (2021) "SPLICE: A Synthetic Paid Loss and Incurred Cost Experience Simulator" \href{https://arxiv.org/abs/2109.04058}{arXiv:2109.04058}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/agi-lab/SPLICE}
  \item Report bugs at \url{https://github.com/agi-lab/SPLICE/issues}
}

}
\author{
\strong{Maintainer}: Melantha Wang \email{wang.melantha@gmail.com}

Authors:
\itemize{
  \item Benjamin Avanzi \email{b.avanzi@unimelb.edu.au}
  \item Greg Taylor \email{greg_taylor60@hotmail.com}
  \item William Ho \email{ho.w@unimelb.edu.au}
}

}
\keyword{internal}
